+++ example/store/user.go
@@ -4,10 +4,10 @@ import "github.com/arstd/light/example/model"
 
 type User interface {
 
-	// insert into users(username, phone, address, status, birthday, created, updated)
-	// values (${u.Username}, ${u.Phone}, ${u.Address}, ${u.Status}, ${u.Birthday},
-	//   CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
-	Insert(u *model.User) (int64, error)
+	// // insert into users(username, phone, address, status, birthday, created, updated)
+	// // values (${u.Username}, ${u.Phone}, ${u.Address}, ${u.Status}, ${u.Birthday},
+	// //   CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
+	// Insert(u *model.User) (int64, error)
 	//
 	// // UPDATE users
 	// // SET [username=${u.Username},]
@@ -25,15 +25,22 @@ type User interface {
 	// // SELECT *
 	// // FROM users WHERE id=${id}
 	// Get(id uint64) (*model.User, error)
-	//
-	// // select *
-	// // from users
-	// // where 1=1
-	// // [and username like ${u.Username}]
-	// // [and phone like ${u.Phone}]
-	// // [and updated > ${u.Updated}]
-	// // limit 20,10
-	// List(u *model.User) ([]*model.User, error)
+
+	// select id, username, phone, address, status, birthday, created, updated
+	// from users
+	// where username like ${u.Username}
+	// [
+	// 	and address = ${u.Address}
+	// 	[and phone like ${u.Phone}]
+	// 	and created > ${u.Created}
+	// ]
+	// and status != ${u.Status}
+	// [and updated > ${u.Updated}]
+	// and birthday is not null
+	// order by updated desc
+	// limit ${offset}, ${size}
+	List(u *model.User, offset, size int) ([]*model.User, error)
+
 	//
 	// // select *
 	// // from users