+++ sqlparser/delete_test.go
@@ -10,7 +10,7 @@ import (
 func TestParseDeleteStmt(t *testing.T) {
 	sql := `DELETE FROM users WHERE id=${id}`
 
-	p := Parser{s: NewScanner(bytes.NewBufferString(sql))}
+	p := NewParser(bytes.NewBufferString(sql))
 	stmt, err := p.Parse()
 	if err != nil {
 		t.Fatal(err)