+++ sqlparser/insert_test.go
@@ -13,7 +13,7 @@ func TestParseInsertStmt(t *testing.T) {
 	values (${u.Username}, ${u.Phone}, ${u.Address}, ${u.Status}, ${u.Birthday},
 	  CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)`
 
-	p := Parser{s: NewScanner(bytes.NewBufferString(sql))}
+	p := NewParser(bytes.NewBufferString(sql))
 	stmt, err := p.Parse()
 	if err != nil {
 		t.Fatal(err)